backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-decouple==3.8
openai>=1.0.0
moviepy==1.0.3
mutagen>=1.47.0
python-multipart==0.0.6
